Simple change, huuuge impact :)

Random numbers from a project I'm working on now:

Before this:
```
# cold load
651137 bytes written (4.87 seconds)

# subsequent bundles
651270 bytes written (2.55 seconds)
```

After the patch:
```
# cold load
651137 bytes written (4.87 seconds)

# subsequent bundles
651270 bytes written (0.53 seconds)
```

It's like 5 times faster already :)