@dariocravero 2016-07-12
Hey @nolanlawson, I don't know if it's possible to get access to npm for this package but if it is I'm happy to push updates with changes like this one or for instance rollup updating versions, etc. :0