@nolanlawson 2016-07-13
So I can't find any documentation on `cache` in Rollup (e.g. on [this page](https://github.com/rollup/rollup/wiki/JavaScript-API)) - can you confirm that it does indeed do a per-module cache, so e.g. if you have multiple bundles, they won't overwrite each other?

As for access to the npm package, yeah I'm happy to make you a contributor and a publisher, but it's still good to get +1s from someone else. :) Typically in my projects I do 1) make a PR, 2) wait for +1, 3) if no response after 24 hours, timebomb and merge it anyway. Cheers!