@dariocravero 2016-07-13
That sounds like a good way to go about it :); I'm dario@uxtemple.com on npm.
As for `cache`, yeah, it's a pity it's not documented :) So far, [rollup takes only the modules from the cache](https://github.com/rollup/rollup/blob/5c0597d70a4a0800bd320d20a229050d73c6daac/src/Bundle.js#L23-L27).
I'm not sure I'm following with multiple bundles, could you elaborate on that? [Do you mean multiple targets](https://github.com/rollup/rollup-watch/blob/master/src/index.js#L106)?