@dariocravero 2016-07-13
(btw, just added some docs around it :P https://github.com/rollup/rollup/wiki/JavaScript-API, [let's hope my assumptions are alright!](https://twitter.com/dariocravero/status/753183526867050496))