@nolanlawson 2016-07-13
Yes, I meant multiple targets. Thanks for the fix to the wiki; makes sense to me. I'll add you as a committer and publisher and publish this as a patch release; cheers!